+++ unreal/AegisWar/Source/AegisWar/Public/WarPlayerController.h
@@ -28,6 +28,8 @@ public:
     UFUNCTION(Server, Reliable) void ServerReturnToDevelopmentSpawn();
     UFUNCTION(Server, Reliable) void ServerEditWorldObject(FName Id, FTransform Transform, bool bObjectHidden, int32 ExpectedRevision);
     UFUNCTION(Server, Reliable) void ServerCreateWorldObject(FName TemplateId, FTransform Transform, int32 ExpectedRevision);
+    UFUNCTION(Server, Reliable) void ServerPlaceWorldObject(FName TemplateId, double Grid, double Angle, int32 ExpectedRevision);
+    UFUNCTION(Server, Reliable) void ServerDropWorldObject(FName Id, int32 ExpectedRevision);
     UFUNCTION(Client, Reliable) void ClientWorldObjectCreated(FName Id);
     UFUNCTION(Server, Reliable) void ServerWorldEditHistory(bool bRedo, int32 ExpectedRevision);
     UFUNCTION(Server, Reliable) void ServerWorldEditDraft(bool bLoad, int32 ExpectedRevision);