+++ unreal/AegisWar/Source/AegisWar/Public/WarWorldEditSubsystem.h
@@ -15,6 +15,8 @@ public:
     bool CanUse(const APlayerController* Controller) const;
     bool Open(APlayerController* Controller, FString& Error);
     bool Create(APlayerController* Controller, FName TemplateId, const FTransform& Transform, int32 Revision, FName& CreatedId, FString& Error);
+    bool CreateInFront(APlayerController* Controller, FName TemplateId, double Grid, double Angle, int32 Revision, FName& CreatedId, FString& Error);
+    bool DropToSurface(APlayerController* Controller, FName Id, int32 Revision, FString& Error);
     bool Edit(APlayerController* Controller, FName Id, const FTransform& Transform, bool bHidden, int32 Revision, FString& Error);
     bool Undo(APlayerController* Controller, bool bRedo, int32 Revision, FString& Error);
     bool SaveDraft(APlayerController* Controller, int32 Revision, FString& Error);